make this project compatible with python3
