+++ utils.py
@@ -1,0 +1,11 @@
+def convert_price_toNumber(price):
+    price = price.split("$")[1]
+    try:
+        price = price.split("\n")[0] + "." + price.split("\n")[1]
+    except:
+        Exception()
+    try:
+        price = price.split(",")[0] + price.split(",")[1]
+    except:
+        Exception()
+    return float(price)
