+++ web_driver_conf.py
@@ -1,0 +1,17 @@
+from selenium import webdriver
+from selenium.webdriver.common.keys import Keys
+
+def get_chrome_web_driver(options):
+    return webdriver.Chrome("./chromedriver", chrome_options=options)
+
+def get_web_driver_options():
+    return webdriver.ChromeOptions()
+
+def set_ignore_certificate_error(options):
+    options.add_argument('--ignore-certificate-errors')
+
+def set_browser_as_incognito(options):
+    options.add_argument('--incognito')
+
+def set_automation_as_head_less(options):
+    options.add_argument('--headless')
